Hask.Constrained

Description

Documentation

newtype p |= q infixr 9

(|=) :: Constraint -> * -> *

This is a corepresentable profunctor

Constructors

Constrained

Fields

runConstrained :: p => q

data EnvC p q

We could flip this around to permit a Curried instance, but then we lose a lot of structure.

Constructors

EnvC (Dict p) q