Hask.Core

Description

This package explores category theory via a couple of non-standard tricks.

First, we use lens-style isomorphism-families to talk about most operations.

Second, we heavily abuse parametricity as a proxy for naturality. This means that the category Nat that gets used throughout is a particularly well-behaved. An inhabitant of Nat :: (i -> j) -> (i -> j) -> * is required to be polymorphic in its argument.

Parametricity is a very strong notion of naturality. Notably, we don't have to care if i or j are co -or- contravariant. (forall a. f a ~> g a) respects _both_ -- at least so long as the functors involved aren't GADTs.

Third, we use kind-indexing to pick the category. This means it is harder to talk about Kleisli categories, etc. but in exchange most of the category selection _just works_. A central working hypothesis of this code is that this is sufficient to talk about interesting categories, and it certainly results in less verbose code than more explicit encodings which clutter up every type class talking about the choice of category.

Here, much of the time the selection is implicit.

Documentation

type family Hom :: i -> i -> k infixr 0

All of our categories will be denoted by the kinds of their arguments.

This is allows for enriched Homs

type (~>) = (Hom :: i -> i -> *) infixr 0

type (^) a b = Internal Hom b a infixr 8

type Dom f = (Hom :: i -> i -> *)

type Cod f = (Hom :: j -> j -> *)

type Cod2 f = (Hom :: k -> k -> *)

type Arr a = (Hom :: i -> i -> *)

type Enriched k = (Hom :: i -> i -> j)

type Internal k = (Hom :: i -> i -> i)

type External k = (Hom :: i -> i -> *)

type Natural k = (Hom :: (i -> j) -> (i -> j) -> *)

Natural transformations

newtype Nat f g infixr 0

Constructors

Nat infixr 0

Fields

runNat :: forall a. f a ~> g a

nat2 :: (forall a b. f a b ~> g a b) -> f ~> g

nat3 :: (forall a b c. f a b c ~> g a b c) -> f ~> g

nat4 :: (forall a b c d. f a b c d ~> g a b c d) -> f ~> g

runNat2 :: Nat (k1 -> k) k2 f g -> Hom k * (f a1 a) (g a1 a)

runNat3 :: Nat (k2 -> k1 -> k) k3 f g -> Hom k * (f a2 a1 a) (g a2 a1 a)

runNat4 :: Nat (k3 -> k2 -> k1 -> k) k4 f g -> Hom k * (f a3 a2 a1 a) (g a3 a2 a1 a)

Functors

class Functor f where

Methods

fmap :: (a ~> b) -> f a ~> f b

first :: Functor f => (a ~> b) -> f a c ~> f b c

class Contravariant f where

Methods

contramap :: (b ~> a) -> f a ~> f b

lmap :: Contravariant f => (a ~> b) -> f b c ~> f a c

Functor composition

class (hom ~ Hom) => Composed hom | k -> hom where

Associated Types

type Compose :: (j -> k) -> (i -> j) -> i -> k

Methods

compose :: f (g a) `hom` Compose f g a

decompose :: Compose f g a `hom` f (g a)

type · = Compose infixr 9

composed :: (Functor k1 k2 (p (Any k)), Contravariant (k1 -> k2) k p, Composed k1 (Hom k1 *), Composed k (Hom k *), Category k2 (Hom k2 *)) => (~>) k2 (p (f (g a)) (f1 (g1 a1))) (p (Compose k k3 k4 f g a) (Compose k1 k5 k6 f1 g1 a1))

associateCompose :: (Functor (k4 -> k3) k (p (Any (k2 -> k1))), Functor k8 k3 f1, Functor k5 k1 f, Contravariant ((k4 -> k3) -> k) (k2 -> k1) p, Composed k8 (Hom k8 *), Composed k5 (Hom k5 *), Composed k3 (Hom k3 *), Composed k1 (Hom k1 *), Category k3 (Hom k3 *), Category k1 (Hom k1 *), Category k (Hom k *)) => (~>) k (p (Compose k1 k5 k2 f (Compose k5 k6 k2 g g1)) (Compose k3 k8 k4 f1 (Compose k8 k7 k4 g3 g2))) (p (Compose k1 k6 k2 (Compose k1 k5 k6 f g) g1) (Compose k3 k7 k4 (Compose k3 k8 k7 f1 g3) g2))

whiskerComposeL :: (Functor k * (Hom k * (Any k)), Contravariant (k -> *) k (Hom k *), Composed k (Hom k *)) => Nat k k2 f g1 -> Nat k k1 (Compose k k2 k1 f g) (Compose k k2 k1 g1 g)

whiskerComposeR :: (Functor x k f, Functor k * (Hom k * (Any k)), Contravariant (k -> *) k (Hom k *), Composed k (Hom k *)) => Nat x k1 g g1 -> Nat k k1 (Compose k x k1 f g) (Compose k x k1 f g1)

lambdaCompose :: (Functor (k4 -> k3) k (p (Any (k2 -> k1))), Contravariant ((k4 -> k3) -> k) (k2 -> k1) p, Composed k3 (Hom k3 *), Composed k1 (Hom k1 *), Identity k3 (Id k3), Identity k1 (Id k1), Category k3 (Hom k3 *), Category k1 (Hom k1 *), Category k (Hom k *)) => (~>) k (p (Compose k1 k1 k2 (Id k1) a) (Compose k3 k3 k4 (Id k3) d)) (p a d)

rhoCompose :: (Functor (k4 -> k3) k (p (Any (k2 -> k1))), Functor k4 k3 c, Functor k2 k1 b, Contravariant ((k4 -> k3) -> k) (k2 -> k1) p, Composed k3 (Hom k3 *), Composed k1 (Hom k1 *), Identity k4 (Id k4), Identity k2 (Id k2), Category k4 (Hom k4 *), Category k3 (Hom k3 *), Category k2 (Hom k2 *), Category k1 (Hom k1 *), Category k (Hom k *)) => (~>) k (p b c) (p (Compose k1 k2 k2 b (Id k2)) (Compose k3 k4 k4 c (Id k4)))

In the limit

class LimC (f `·` p) => Post f p

Derived notions

fmap1 :: forall p a b c. Post Functor p => (a ~> b) -> p c a ~> p c b

contramap1 :: forall p a b c. Post Contravariant p => (a ~> b) -> p c b ~> p c a

fmap2 :: forall p a b c d. Post (Post Functor) p => (a ~> b) -> p c d a ~> p c d b

class (Functor p, Post Functor p) => Bifunctor p

bimap :: (Bifunctor p, Category (Cod2 p)) => (a ~> b) -> (c ~> d) -> p a c ~> p b d

class (Contravariant p, Post Functor p) => Profunctor p

dimap :: (Profunctor p, Category (Cod2 p)) => (a ~> b) -> (c ~> d) -> p b c ~> p a d

class (Contravariant p, Functor p) => Phantom p

Colim -| Const -| Lim

class (k ~ (~>)) => Cocomplete k where

Minimal complete definition

colim

Associated Types

type Colim :: (i -> j) -> j

Methods

colim :: k (f a) (Colim f)

cocomplete :: Dict ((Colim :: (i -> j) -> j) -| Const)

class (Const ~ k) => Constant k | j i -> k where

Minimal complete definition

Nothing

Associated Types

type Const :: j -> i -> j

Methods

_Const :: Iso (k b a) (k b' a') b b'

class (hom ~ Hom) => Complete hom | j -> hom where

A complete category has all small limits.

Minimal complete definition

elim

Associated Types

type Lim :: (i -> j) -> j

Methods

elim :: hom (Lim g) (g a)

complete :: Dict (Const -| (Lim :: (i -> j) -> j))

Natural Isomorphisms

type Iso s t a b = forall p. Profunctor p => p a b -> p s t

type Iso' s a = Iso s s a a

Adjunctions

class (Functor f, Functor u) => f -| u | f -> u, u -> f where infixr 0

f -| u indicates f is left adjoint to u

Methods

adj :: Iso (f a ~> b) (f a' ~> b') (a ~> u b) (a' ~> u b')

unitAdj :: (f -| u, Category (Dom u)) => a ~> u (f a)

counitAdj :: (f -| u, Category (Dom f)) => f (u b) ~> b

zipR :: (f -| u, Cartesian (Dom u), Cartesian (Cod u)) => Iso (u a * u b) (u a' * u b') (u (a * b)) (u (a' * b'))

absurdL :: (f -| u, Initial z) => Iso z z (f z) (f z)

cozipL :: (f -| u, Cocartesian (Cod u), Cocartesian (Cod f)) => Iso (f (a + b)) (f (a' + b')) (f a + f b) (f a' + f b')

Currying

class Curried p e | p -> e, e -> p where

Minimal complete definition

curry, uncurry | apply, unapply

Associated Types

type Curryable p :: k -> Constraint

Methods

curry :: Curryable p a => (p a b ~> c) -> a ~> e b c

uncurry :: (a ~> e b c) -> p a b ~> c

apply :: p (e a b) a ~> b

unapply :: Curryable p a => a ~> e b (p a b)

curried :: (Curried p e, Curryable p a) => Iso (p a b ~> c) (p a' b' ~> c') (a ~> e b c) (a' ~> e b' c')

ccc :: (Category (Cod2 p), Symmetric p, Curried p e, Curryable p a) => Iso (p i a ~> b) (p i' a' ~> b') (a ~> e i b) (a' ~> e i' b')

Cocurrying

class Cocurried f u | f -> u, u -> f where

Associated Types

type Cocurryable u :: k -> Constraint

Methods

cocurry :: (f a b ~> c) -> b ~> u c a

uncocurry :: Cocurryable u c => (b ~> u c a) -> f a b ~> c

cocurried :: (Cocurried f u, Cocurryable u c') => Iso (f a b ~> c) (f a' b' ~> c') (b ~> u c a) (b' ~> u c' a')

Automatic Lifting

class (hom ~ Hom) => Lifted hom where

Minimal complete definition

Nothing

Associated Types

type Lift :: (j -> k -> l) -> (i -> j) -> (i -> k) -> i -> l

Methods

_Lift :: forall q r f h g e a b. Iso (Lift q f g a) (Lift r h e b) (q (f a) (g a)) (r (h b) (e b))

Constraints

newtype a :- b :: Constraint -> Constraint -> *

Constructors

Sub (a -> Dict b)

data Dict $a :: Constraint -> * where

Constructors

Dict :: a => Dict a

(\\) :: a => (b -> r) -> (:-) a b -> r

class (p, q) => p & q infixr 2

class p |- q where infixr 9

Methods

implies :: p :- q

_Sub :: Iso (p :- q) (p' :- q') (Dict p -> Dict q) (Dict p' -> Dict q')

_Implies :: Iso (p :- q) (p' :- q') (Dict (p |- q)) (Dict (p' |- q'))

Lens Esoterica

The Yoneda embedding

newtype Get r a b

Get is the Yoneda embedding C -> [ C^op, Set]

How? We can view it as a functor from C -> [ C^op X 1, Set ] with the latter embedded in Prof.

Where does the 1 come from? We use the existence of Contravariance and Covariance in all arguments to make a category where every object is the same as long as the kind chosen has either an initial or a terminal object. But get is polymorphic in the choice of kind for the b parameter, making it possible to choose it to be anything you want.

Constructors

Get

Fields

runGet :: a ~> r

_Get :: (Functor * k (p (Any *)), Contravariant (* -> k) * p, Category k (Hom k *)) => (~>) k (p (Hom k2 * a r) (Hom k4 * a1 r1)) (p (Get k1 k2 r a b) (Get k3 k4 r1 a1 b1))

get :: (Category c, c ~ (~>)) => (Get a a a -> Get a s s) -> c s a

newtype Beget r a b

Beget is the contravariant Yoneda embedding. C^op -> [C, Set]

we can view it as a functor from C^op -> [ 1^op X C, Set ] with the latter embedded in Prof.

Constructors

Beget

Fields

runBeget :: r ~> b

_Beget :: (Functor * k (p (Any *)), Contravariant (* -> k) * p, Category k (Hom k *)) => (~>) k (p (Hom k2 * r b) (Hom k4 * r1 b1)) (p (Beget k1 k2 r a b) (Beget k3 k4 r1 a1 b1))

beget :: (Category c, c ~ (~>)) => (Beget b b b -> Beget b t t) -> c b t

(#) :: (Beget b b b -> Beget b t t) -> b -> t

Inverting natural isomorphisms

newtype Un p a b s t

Constructors

Un

Fields

runUn :: p t s ~> p b a

_Un :: (Functor * k (p (Any *)), Contravariant (* -> k) * p, Category k (Hom k *)) => (~>) k (p (Hom k1 * (p1 t s) (p1 b a)) (Hom k2 * (p2 t1 s1) (p2 b1 a1))) (p (Un i k1 p1 a b s t) (Un i1 k2 p2 a1 b1 s1 t1))

un :: (Un p a b a b -> Un p a b s t) -> p t s -> p b a

data Via a b s t where

Constructors

Via :: (s ~> a) -> (b ~> t) -> Via a b s t

via :: forall a b r p c t u. Category p => (Via a b a b -> Via r (p c c) u t) -> (t -> u) -> r

get   = via _Get
beget = via _Beget
un    = via _Un

mapping :: (Functor f, Functor f', Category (Dom f)) => (Via a b a b -> Via a b s t) -> Iso (f s) (f' t) (f a) (f' b)

lmapping :: (Contravariant f, Contravariant f', Category (Dom f)) => (Via s t s t -> Via s t a b) -> Iso (f s x) (f' t y) (f a x) (f' b y)

swapping :: (Profunctor f, Profunctor f', Category (Dom f), Category (Cod2 f), Category (Cod2 f')) => (Via a b a b -> Via a b s t) -> Iso (f a s) (f' b t) (f s a) (f' t b)

firstly :: (Functor f, Functor f', Category (Dom f)) => (Via a b a b -> Via a b s t) -> Iso (f s x) (f' t y) (f a x) (f' b y)

post :: Category (Arr i) => (Via a b a b -> Via a b s t) -> Iso (i ~> s) (j ~> t) (i ~> a) (j ~> b)

pre :: Category (Arr i) => (Via a b a b -> Via a b s t) -> Iso (a ~> i) (b ~> j) (s ~> i) (t ~> j)

Hom as a Profunctor

newtype Self a b

Constructors

Self

Fields

runSelf :: a ~> b

_Self :: (Functor * k (p (Any *)), Contravariant (* -> k) * p, Category k (Hom k *)) => (~>) k (p (Hom k1 * a b) (Hom k2 * a1 b1)) (p (Self k1 a b) (Self k2 a1 b1))

Monoidal Categories

class (Bifunctor p, Profunctor (Cod2 p), Category (Cod2 p)) => Semitensor p where

Methods

associate :: Iso (p (p a b) c) (p (p a' b') c') (p a (p b c)) (p a' (p b' c'))

type family I p :: x

class Semitensor p => Tensor p where

Associated Types

type Tensorable p :: x -> Constraint

Methods

lambda :: Tensorable p a => Iso (p (I p) a) (p (I p) a) a a

rho :: Tensorable p a => Iso (p a (I p)) (p a (I p)) a a

class Bifunctor p => Symmetric p where

Methods

swap :: p a b ~> p b a

Closed Categories

class (Profunctor (Internal k), Category k, k ~ Hom) => Closed k where

Minimal complete definition

Nothing

Methods

iota :: Iso (Internal k (I (Internal k)) a) (Internal k (I (Internal k)) b) a b

jot :: I (Internal k) ~> Internal k a a

Terminal and Initial Objects

class (One ~ t) => Terminal t | i -> t where

Associated Types

type One :: i

Methods

terminal :: a ~> t

class (Zero ~ t) => Initial t | i -> t where

Associated Types

type Zero :: i

Methods

initial :: t ~> a

Copowers and Cartesian Categories

type family Copower :: i -> j -> j

This is factored out of Precartesian so that we can also use it for copowers/tensors

type * = (Copower :: i -> i -> i) infixl 7

class (h ~ (~>), Symmetric (* :: i -> i -> i), Semitensor (* :: i -> i -> i)) => Precartesian h | i -> h where

Methods

fst :: forall a b. (a * b) ~> a

snd :: forall a b. (a * b) ~> b

(&&&) :: forall a b c. (a ~> b) -> (a ~> c) -> a ~> (b * c)

class (h ~ (~>), Tensor (* :: i -> i -> i), Terminal (I (* :: i -> i -> i)), Precartesian h) => Cartesian h | i -> h

class (h ~ (~>), Symmetric ((+) :: i -> i -> i), Semitensor ((+) :: i -> i -> i)) => Precocartesian h | i -> h where

Associated Types

type (+) :: i -> i -> i infixl 6

Methods

inl :: forall a b. a ~> (a + b)

inr :: forall a b. b ~> (a + b)

(|||) :: forall a b c. (a ~> c) -> (b ~> c) -> (a + b) ~> c

class (h ~ (~>), Tensor ((+) :: i -> i -> i), Initial (I ((+) :: i -> i -> i)), Precocartesian h) => Cocartesian h | i -> h

Distributive bicartesian categories

class (Cartesian k, Cocartesian k) => Distributive k | i -> k where

Methods

distribute :: forall a b c a' b' c'. Iso ((a * b) + (a * c)) ((a' * b') + (a' * c')) (a * (b + c)) (a' * (b' + c'))

CCC's

class (Cartesian k, Closed k, Curried * (Internal k), I (Internal k) ~ I *, Curryable * (I (Internal k))) => CCC k | x -> k

Monoidal Functors

class (Precartesian ((~>) :: x -> x -> *), Precartesian ((~>) :: y -> y -> *), Functor f) => Semimonoidal f where

Methods

ap2 :: (f a * f b) ~> f (a * b)

ap :: (Semimonoidal f, CCC (Dom f), CCC (Cod f), Curryable * (f (b ^ a))) => f (b ^ a) ~> (f b ^ f a)

ap2Applicative :: Applicative f => (f a * f b) -> f (a * b)

class (Cartesian ((~>) :: x -> x -> *), Cartesian ((~>) :: y -> y -> *), Semimonoidal f) => Monoidal f where

Methods

ap0 :: One ~> f One

return :: (Monoidal f, Strength f, CCC (Dom f), Tensorable * a) => a ~> f a

class (Precocartesian ((~>) :: x -> x -> *), Precocartesian ((~>) :: y -> y -> *), Functor f) => Cosemimonoidal f where

Methods

op2 :: f (a + b) ~> (f a + f b)

class (Cocartesian ((~>) :: x -> x -> *), Cocartesian ((~>) :: y -> y -> *), Cosemimonoidal f) => Comonoidal f where

Methods

op0 :: f Zero ~> Zero

Monads

class Semimonoidal (m :: x -> x) => Semimonad m where

Minimal complete definition

join | bind

Methods

join :: m (m a) ~> m a

bind :: Semimonad m => (a ~> m b) ~> (m a ~> m b)

class (Monoidal m, Semimonad m) => Monad m

class (Functor f, Category (Dom f)) => Cosemimonad f where

Minimal complete definition

duplicate | extend

Methods

duplicate :: f a ~> f (f a)

extend :: (f a ~> b) -> f a ~> f b

class Cosemimonad f => Comonad f where

Methods

extract :: f a ~> a

Monoids

class Precartesian (Arr m) => Semigroup m where

Methods

mult :: (m * m) ~> m

multM :: (Semimonoidal u, Semigroup m) => (u m * u m) ~> u m

class (Semigroup m, Cartesian (Arr m)) => Monoid m where

Methods

one :: One ~> m

oneM :: (Monoidal u, Monoid m) => One ~> u m

mappend :: (Semigroup m, CCC (Arr m), Curryable * m) => m ~> (m ^ m)

class Precocartesian (Arr m) => Cosemigroup m where

Methods

comult :: m ~> (m + m)

zeroOp :: (Comonoidal f, Comonoid m) => f m ~> Zero

class (Cosemigroup m, Cocartesian (Arr m)) => Comonoid m where

Methods

zero :: m ~> Zero

comultOp :: (Cosemimonoidal f, Cosemigroup m) => f m ~> (f m + f m)

Strength

class Functor f => Strength f where

Methods

strength :: (a * f b) ~> f (a * b)

class Functor f => Costrength f where

Methods

costrength :: f (a + b) ~> (a + f b)

Identity functors

class (f -| f, Id ~ f) => Identity f | i -> f where

Minimal complete definition

Nothing

Associated Types

type Id :: i -> i

Methods

_Id :: Iso (f a) (f a') a a'

Kan extensions

class (c ~ Hom) => HasRan c | k -> c where

Minimal complete definition

iotaRan, jotRan

Associated Types

type Ran :: (i -> j) -> (i -> k) -> j -> k

Methods

ranCurried :: Dict (Curried Compose (Ran :: (i -> j) -> (i -> k) -> j -> k))

ranProfunctor :: Category (Hom :: j -> j -> *) => Dict (Profunctor (Ran :: (i -> j) -> (i -> k) -> j -> k))

ranFunctor :: Dict (Functor (Ran f g :: j -> k))

iotaRan :: forall f g id. (Category (Cod f), Category (Dom f), Functor g, Identity id) => Iso (Ran id f) (Ran id g) f g

jotRan :: forall f. Id ~> Ran f f

return for Codensity

class (c ~ Hom) => HasLan c | k -> c where

Minimal complete definition

epsilonLan

Associated Types

type Lan :: (i -> j) -> (i -> k) -> j -> k

Methods

lanCocurried :: Dict (Cocurried (Lan :: (i -> j) -> (i -> k) -> j -> k) Compose)

lanProfunctor :: Category (Hom :: j -> j -> *) => Dict (Profunctor (Lan :: (i -> j) -> (i -> k) -> j -> k))

lanFunctor :: Dict (Functor (Lan f g :: j -> k))

epsilonLan :: forall f g id. (Category (Cod f), Category (Dom f), Functor f, Identity id) => Iso (Lan id f) (Lan id g) f g

Categories

Products

data Prod :: (i, j) -> (i, j) -> * where

Constructors

Want :: Prod a a

Have :: (a ~> b) -> (c ~> d) -> Prod `(a, c)` `(b, d)`

runProd :: forall a b c d. (Category ((~>) :: i -> i -> *), Category ((~>) :: j -> j -> *)) => Prod `(a, c)` `(b, d)` -> (a ~> b, c ~> d)

runFst :: forall a b c d. Category ((~>) :: i -> i -> *) => Prod `(a, c)` `(b, d)` -> a ~> b

runSnd :: forall a b c d. Category ((~>) :: j -> j -> *) => Prod `(a, c)` `(b, d)` -> c ~> d

diagProdAdj :: forall a b c a' b' c'. Cartesian ((~>) :: i -> i -> *) => Iso (`(a, a)` ~> `(b, c)`) (`(a', a')` ~> `(b', c')`) (a ~> (b * c)) (a' ~> (b' * c'))

sumDiagAdj :: forall a b c a' b' c'. Cocartesian ((~>) :: i -> i -> *) => Iso ((b + c) ~> a) ((b' + c') ~> a') (`(b, c)` ~> `(a, a)`) (`(b', c')` ~> `(a', a')`)

Unit

data Unit a b

Constructors

Unit

Empty

newtype Empty a b

Constructors

Empty (Empty a b)

data No a

Candidates for Removal:

class (Profunctor p, p ~ (~>)) => Semicategory p

Internals

Categories

Composition

newtype Compose1 f g a

Constructors

Compose

Fields

getCompose :: f (g a)

newtype Compose2 f g a b

Constructors

Compose2

Fields

getCompose2 :: f (g a) b

class f (g a) => ComposeC f g a

We can compose constraints

Limits

newtype Lim1 f

Constructors

Lim

Fields

getLim :: forall x. f x

newtype Lim2 f y

Constructors

Lim2

Fields

getLim2 :: forall x. f x y

newtype Lim3 f y z

Constructors

Lim3

Fields

getLim3 :: forall x. f x y z

class LimC p where

LimC :: (i -> Constraint) -> Constraint takes limits in the category of constraints

type instance Lim = LimC

Methods

limDict :: Dict (p a)

newtype Const1 b a

Constructors

Const

Fields

getConst :: b

newtype Const2 f a c

Constructors

Const2

Fields

getConst2 :: f c

class b => ConstC b a

data Colim1 f where

Constructors

Colim :: f x -> Colim1 f

data Colim2 f x where

Constructors

Colim2 :: f y x -> Colim2 f x

Identities

newtype Id0 a

Constructors

Id

Fields

runId :: a

newtype Id1 f a

Constructors

Id1

Fields

runId1 :: f a

class c => IdC c

Lifts

newtype Lift1 p f g a

Constructors

Lift

Fields

lower :: p (f a) (g a)

newtype Lift2 p f g a b

Constructors

Lift2

Fields

lower2 :: p (f a) (g a) b

class r (p a) (q a) => LiftC r p q a

Copowers

data Copower1 x f a

Constructors

Copower x (f a)

data Copower2 x f a b

Constructors

Copower2 x (f a b)

data Copower2_1 x f a b

Constructors

Copower2_1 (x a) (f a b)

Kan extensions

data Ran1 f g a

Constructors

forall z . Functor z => Ran (Compose z f ~> g) (z a)

newtype Lan1 f g a

Constructors

Lan

Fields

runLan :: forall z. Functor z => (g ~> Compose z f) ~> z a

newtype Lan2 f g a b

Constructors

Lan2

Fields

runLan2 :: forall z. Functor z => (g ~> Compose z f) ~> z a b

Base Re-exports

data Constraint :: BOX

Prelude Re-Exports

undefined :: a

A special case of error. It is expected that compilers will recognize this and insert error messages which are more appropriate to the context in which undefined appears.

($) :: (a -> b) -> a -> b infixr 0

Application operator. This operator is redundant, since ordinary application (f x) means the same as (f $ x). However, $ has low, right-associative binding precedence, so it sometimes allows parentheses to be omitted; for example:

    f $ g $ h x  =  f (g (h x))

It is also useful in higher-order situations, such as map ($ 0) xs, or zipWith ($) fs xs.

data Either a b :: * -> * -> *

The Either type represents values with two possibilities: a value of type Either a b is either Left a or Right b.

The Either type is sometimes used to represent a value which is either correct or an error; by convention, the Left constructor is used to hold an error value and the Right constructor is used to hold a correct value (mnemonic: "right" also means "correct").

Constructors

Left a

Right b

data Maybe a :: * -> *

The Maybe type encapsulates an optional value. A value of type Maybe a either contains a value of type a (represented as Just a), or it is empty (represented as Nothing). Using Maybe is a good way to deal with errors or exceptional cases without resorting to drastic measures such as error.

The Maybe type is also a monad. It is a simple kind of error monad, where all errors are represented by Nothing. A richer error monad can be built using the Either type.

Constructors

Nothing

Just a

Re-exports

newtype Tagged s b :: k -> * -> *

Constructors

Tagged

Fields

unTagged :: b

data Proxy t :: k -> *

A concrete, poly-kinded proxy type

Constructors

Proxy

class Category cat where

A class for categories. id and (.) must form a monoid.

Methods

id :: cat a a

the identity morphism

(.) :: cat b c -> cat a b -> cat a c infixr 9

morphism composition

data Void :: *

absurd :: Void -> a