Hask.Rel

Documentation

type family Rel :: i -> j

Rel should be a faithful functor that is also injective on objects.

and it sometimes has adjoints, if either exist, then they should satisfy the following relationship

Reflected -| Rel -| Coreflected

This provides the inclusion functor for a subcategory with objects indexed in i into a category with objects in j

type family Reflected :: j -> i

The "reflector" of a reflective sub-category with objects in j of a category with objects in i.

Reflected -| Rel

type family Coreflected :: j -> i

This "coreflector" of a coreflective sub-category with objects in j of a category with objects in i.

Rel -| Coreflected

class f ~| u | f k -> u, u j -> f where

Relative adjoints

Methods

radj :: Iso (f a ~> b) (f a' ~> b') (Rel a ~> u b) (Rel a' ~> u b')

When Rel = Identity, a relative adjunction is just a normal adjunction, and you can use:

radj = adj.pre _Id

class Functor m => RelMonad m where

Relative monads

Methods

runit :: Rel a ~> m a

rbind :: (Rel a ~> m b) -> m a ~> m b

class (rel ~ Rel, Tensor (RelCompose :: (j -> c) -> (j -> c) -> j -> c)) => RelComposed rel where

Associated Types

type RelCompose :: (j -> c) -> (j -> c) -> j -> c

Methods

rcomposed :: Functor f' => Iso (RelCompose f) (RelCompose f') (Compose (Lan rel f)) (Compose (Lan rel f'))

newtype ComposeConst1 f g a b

Constructors

ComposeConst1

Fields

decomposeConst1 :: Compose2 (Lan2 Const1 f) g a b